[license][1/2] Automatically update the BSL license's change date
Why
---
We want to keep the BSL license's change date to be three years after each code commit. It is nicer to have this automated. This is part 1, which doesn't push to `main` yet.

How
---
Add a GitHub Action workflow that is set to run daily. It compares the commit time (not author time) of the last commit that modified a file other than LICENSE with the current change date, taking the three year period into account.

Test Plan
---
Add a new workflow with:
```
on:
  pull_request
```
so that we can test it instead of relying on the daily cron job. After committing this, we will also be able to verify that it runs on the cron schedule.

The expected output in the GHA logs is for `git show` to show a commit that updates the license's date.