[license] Automatically update the BSL license's change date (#98)
[license] Automatically update the BSL license's change date

Why
---
We want to keep the BSL license's change date to be three years after each code commit. It is nicer to have this automated.

Removed the script that ran during publishing.

How
---
Add a GitHub Action workflow that is set to run every three months (though for testing, it's set to daily). It compares the "change date" in the license file with the date 3 years in the future and updates the license file if needed.

Test Plan
---
Add a new workflow with:
```
on:
  pull_request
```
so that we can test it instead of relying on the daily cron job. After committing this, we will also be able to verify that it runs on the cron schedule.

The expected output in the GHA logs is for `git show` to show a commit that updates the license's date.