[1/1] Support resource-flag in eas-build-job
# Why
In order to support resource flags on EAS Build commands, we need to
make some modifications to the eas-build-job package.

# How
Add a new enum for BuildResourceClass that matches the enum created
for GraphQL in universe/server/www. Then use that enum to specify new
buildResourceClass attributes on JobSchema objects for both Android and
iOS.

# Test Plan
```
yarn test
```
See test plan on this PR for more thorough e2e testing summary:
https://github.com/expo/universe/pull/9800