[logger] Add callback
Added an optional callback to the function piping spawn output into the child logger. This callback gets executed whenever the child logger calls the info() function to print new log line. In this case it is needed to refresh a timeout so it gets reset after each line from spawn process

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install