[turtle-spawn] Add and pass callback
Added an optional callback to the spawn function. This callback gets passed on to the pipeSpawnOutput function to be executed on each line printed by the logger from the spawn process. In this case it is needed to refresh a timeout so it gets reset after each line from spawn process

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install