[logger] [ENG-10483] Add callback (#298)
* [logger] Add callback

Added an optional callback to the function piping spawn output into the child logger. This callback gets executed whenever the child logger calls the info() function to print new log line. In this case it is needed to refresh a timeout so it gets reset after each line from spawn process

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [logger] Simplify

Made the callback call as a one-liner

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [logger] Extend options

Passed callback as a part of PipeOptions

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [logger] Export PipeOptions

Exported PipeOptions to use in `turtle-spawn`

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install