[turtle-spawn] [ENG-10483] Add and pass callback (#299)
* [turtle-spawn] Add and pass callback

Added an optional callback to the spawn function. This callback gets passed on to the pipeSpawnOutput function to be executed on each line printed by the logger from the spawn process. In this case it is needed to refresh a timeout so it gets reset after each line from spawn process

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [turtle-spawn] Don't redefine options

Imported PipeOptions from logger and used them to define the SpawnOptions instead of redefining them manually

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install