[build-tools] [ENG-10483] Add timeouts for install dependencies phase (#297)
* [build-tools] Add timeouts

Adds two timeouts to the install dependencies phase. First executes after 15 minutes and leaves a warning message in the logs for the user to check their package.json. Second executes after 30 minutes leaving an error message and throwing an error. If phase ends before the timeouts they are cleared

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [build-tools] [wip] Use Promise.race()

Used Promise.race() to avoid restarting the build after an error. Still not ideal solution

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [build-tools] Kill spawn process

Used similar method to the timeout on Expo Doctor phase, seems to be working ok

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [build-tools] Refresh timeouts after each line

Passed a callback to the spawn process refreshing the timeouts after each line. This way the warning message and the kill error happen after a period of inactivity instead of a flat period after start of the phase

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [build-tools] Don't redefine options

Imported SpawnOptions from turtle-spawn and used them to define the arguments in installDependenciesAsync() instead of redefining them manually

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install

* [build-tools] Match argument names

Renamed arguments to match the ones defined in SpawnOptions

See: https://linear.app/expo/issue/ENG-10483/investigate-why-a-pair-of-builds-froze-during-yarn-install