Create `eas/start_ios_simulator` build function (#323)
# Why

Part of the Maestro example.

# How

Moved my tested EAS workflow into a TS function.

It prints all available simulator in the beginning so the users have a list to compare their request to, if they need to. We'll need to make sure to educate people that Simulator Devices, UDIDs are image-specific.

A better version of this function could accept an iOS version to test on, would download and install necessary runtime and create a dedicated device for tests, but I suspect this is enough for now.

# Test Plan

I tested it a bunch, with Simulator stopped, started, with different inputs (valid, invalid).

One thing I don't like is that stdout and stderr of a `spawn` can lose ordering, see:
<img width="1023" alt="Zrzut ekranu 2024-01-26 o 22 14 51" src="https://github.com/expo/eas-build/assets/1151041/0034f89e-612e-48c9-aa75-38f9cf2baeee">
<img width="979" alt="Zrzut ekranu 2024-01-26 o 22 14 38" src="https://github.com/expo/eas-build/assets/1151041/4c897262-1466-4f17-aca9-8e9b076f6276">

Is there anything we can do about it? Maybe I'm doing something wrong?