Support glob in artifact path(s) (#337)
# Why

I was trying to add an example for uploading multiple artifacts — an IPA and an APP. I wanted to use `upload-artifact` to grab the APP, so I tried using one of https://github.com/expo/eas-build/blob/61fcb7d3662c8964da28272ca6c9af5626cb42fb/packages/build-tools/src/steps/functions/findAndUploadBuildArtifacts.ts#L96-L103 as path. As glob was not supported, this was not possible.

# How

As we call `findAndUploadArtifacts` before handing artifact off to `ctx.runtimeApi.uploadArtifact`, we'll do the same in `uploadArtifacts`.

Adds support for `paths` input which is expected to be an array of strings (as validated by Zod).

# Test Plan

Added an automated test.