Add `eas/resolve_build_config` function (#364)
# Why

GitHub build job data is missing a lot of properties, because we don't provide them when we create the build, because we can't, because the config may need to be resolved dynamically, with external dependencies etc.

For this reason `eas build:internal` exists which we call during the build, once all dependencies are installed. We set this up in `get_credentials_for_build_triggered_by_github_integration`. This covered half of the use cases. Turns out we need to call this also in builds that don't need credentials so they have `buildType`, `version` and other properties.

# How

Added `eas/resolve_build_config` function that will be called in _all_ `eas/build` flows. If the build has been triggered by a GitHub trigger, it will call `eas build:internal` to resolve necessary properties, update build information in `www` and in context. It will always print the configuration (with `secrets` and `projectArchive` omitted, as we omit it in the regular build process).

I'm not sure if printing all or even some job properties makes sense. For me, the advantage is that a user can inspect what is the configuration, what properties can they access in `${ eas.job. ... }`. On the other hand it feels noisy. What is your opinion?

# Test Plan

<img width="626" alt="Zrzut ekranu 2024-03-13 o 22 22 41" src="https://github.com/expo/eas-build/assets/1151041/bc0ed085-d88b-4c24-a7e9-f23b37744fa1">