Export `runGenericJobAsync` (#373)
# Why

We'll use it to trigger generic jobs. For a second I thought we could reuse `runCustomBuildAsync`, but then I figured maybe it's better to have something separate.

Also, wraps preparing project with a build phase. We don't have this in custom build. I think it makes sense to track the time of archive / Git repository download? Or maybe not?

# How

Copied `runCustomBuildAsync`, removed a bunch of things. Reverted changes to `runCustomBuildAsync` done recently.

# Test Plan

This adds something new, so should be ok.