Add workflow for "EAS Update in the Cloud" (#378)
# Why

We want to enable "internal workflows" -- jobs that we maintain the workflow of. This will come in handy now for EAS Update in the Cloud, where we can trigger a job with `path: __eas/eas-update-in-the-cloyd.yml` and _maybe_ can let us replace legacy build flows some day.

# How

This feels weird, but is actually quite smart.

By having workflow YAMLs like this, we:
- can point people here so if they need to customize it they can copy and customize,
- can show people exactly what we do, no black box magic,
- we dogfood workflow YAML
- we are not introducing any special branching in code, it's still YAML in a file.

I'm not a fan of how we're adding the workflows to the job. Depending on `__dirname` and relative path to `resources` is sooo brittle. However, I can't seem to find a better way of making it more strict without a bigger overhaul of the way we build `worker` (we would need to introduce a compiler/bundler capable of understanding that when we import a `.yml` we want to read it as a string, etc.).

# Test Plan

I was able to trigger a test job run by setting custom config path to `__eas/eas-update-in-the-cloud.yml`.