[build-tools] Add EXPO_TOKEN to eas-cli call (#390)
# Why

We need to expose the `robotAccessToken` if we want to use it.

# How

Used `command` interpolation. From what I could see we don't interpolate values in `env`. I checked that when interpolating, undefined falls back to `''`, so this should not break development (where we won't have `robotAccessToken` and instead rely on already authenticated `eas-cli`).

# Test Plan

Will test on staging.