Require both `gitCommitHash` and `gitRef` (#395)
# Why

Some cleanup after the migration period.

# How

Made both `gitRef` and `gitCommitHash` required (as in non-optional). `gitCommitHash` is expected to always be provided, `gitRef` can be `null` (if we're building not from a branch or a tag, but from a detached commit). We are not going to use the `gitRef = null` case in `www` yet.

# Test Plan

I built this package and linked to `www` in https://github.com/expo/universe/pull/15254. The changes required to make `tsc` ok seemed forward-compatible.