[steps] [ENG-11679] Fix escape characters in step inputs (#408)
* [steps] Fix escape characters

The input value is JSON.stringified for serialization, and therefore needs to JSON.parsed to remove the doubly escaped characters. This fixes all escape character, not only `\n` as before

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [steps] Fix escape characters in object interpolation

The string values had to be fixed too when the object gets interpolated with them

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [steps] Fix syntaxError

The parsing of interpolated strings in objects is not needed, since the properties get later parsed anyway and doubly parsed properties were the reason for syntaxErrors. Removed the broken test too

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [steps] Fix test

The test was failing after recent change

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [build-tools] Fix tests

The tests were failing after recent change

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [build-tools] [steps] Re-add fallback

Applied again the fallback for when the input is not JSON.stringified, if it ever happens, and the raw string is returned in such a case

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds

* [build-tools] [steps] Re-add fallback

Applied again the fallback for when the input is not JSON.stringified, if it ever happens, and the raw string is returned in such a case

See: https://linear.app/expo/issue/ENG-11679/investigate-incorrect-escaping-of-special-characters-in-custom-builds