[eas-build] [steps] [ENG-13151] Catch errors while evaluating if step should be executed (#436)
* [build-tools] [steps] Catch errors when evaluating if step should be executed

Wrapped the function determining if the step should be executed in a try/catch block. The caught error is logged, as is a message that the runner failed to evaluate if the step should be executed or not, so the user knows what's going on. The step is then skipped and the build is failed

See: https://linear.app/expo/issue/ENG-13151/commented-out-lines-in-message-template-fail-the-send-slack-message

* Update packages/steps/src/BuildWorkflow.ts

Co-authored-by: Szymon Dziedzic <szymon.dziedzic@swmansion.com>

---------

Co-authored-by: Szymon Dziedzic <szymon.dziedzic@swmansion.com>