[eas-build] Expose generic job's `BuildWorkflow` (#444)
# Why

Exposing `BuildWorkflow` from `runGenericJobAsync` lets us fetch individual steps and their outputs after the job is complete.

# How

- wraps `buildWorkflow.executeAsync()` with `asyncResult` and returns the result from `runGenericJobAsync` alongside `buildWorkflow`,
- adds `outputs` to `Generic.Job` so the user has a way of telling us which steps outputs are to be considered jobs outputs,
- when a generic job finishes, Worker sends a `robotAccessToken`-authenticated request to [a new `www` endpoint](https://github.com/expo/universe/pull/16916/files).

Also changes the way we're handling `outputsDir` and `envsDir` — we want the outputs and environment variables to be collected always, not only if the step succeeded. And exposes `jsepEval` so we can use it in Worker.

Moreover, removes `outputs` from `BuildStep`. We had two: `outputs` and `outputById`, one being updated as we process the workflow and the other not.

# Test Plan

Adjusted tests. I've used this version to run a custom job with `set-output name test` and gotten 

<img width="430" alt="Zrzut ekranu 2024-09-27 o 18 44 42" src="https://github.com/user-attachments/assets/c0b53548-b913-454e-b077-77e071445aa0">