[build-tools] Do not let eas-cli override initiatingUserId or appId (#454)
# Why

Right now `eas-cli` is not returning `initiatingUserId` or `appId` because we haven't finished https://github.com/expo/eas-cli/pull/2652.

Actually, this is good, because we don't need to rely on `eas-cli` on the topic of `initiatingUserId` or `appId`. Whatever `www` has set on the original job should stand even after build config resolution.

# How

Retained `appId` and `initiatingUserId` on the `sanitizeJob` call.

# Test Plan

Turtle should deploy with this change.