[build-tools] automatially parse Xcode logs in search for 'error: ' logs to make debugging build failures easier for user (#471)
# Why

Talked about it with @lukmccall today. We want to improve our old generic error message which we display when the build fails during the `RUN_FASTLANE` phase that says `The "Run fastlane" step failed with an unknown error. Refer to "Xcode Logs" below for additional, more detailed logs.` to make it more informative by parsing Xcode logs for the user and displaying found errors in the error message itself so it is easier to find them.

# How

Add a new user error handler for `XCODE_BUILD_PHASE`. If this error handler finds `error: ...` logs in the Xcode build logs file, it displays them nicely in the user-facing error message.

Similar to: https://github.com/expo/eas-build/pull/148

# Test Plan

Added automated tests