[eas-build-job] Allow `steps` in build jobs (#493)
# Why

For workflows it'll be much easier to let users define `steps` in the workflow file instead of having to do the custom build config and stuff.

# How

Added ability to pass `steps` in build jobs. Alongside, we need to send `workflowInteprolationContext`, so added that too.

# Test Plan

Tested manually by running both manual and GitHub workflow build jobs. Also ran a custom build and a regular build from `eas-cli`.