[steps] Do not trim step output value (#502)
# Why

I think we've been doing `.trim()` on output values because of plain `echo` in `set-output`.

# How

Added `-n` to `echo`, removed `.trim()`. Then changed to use Base64 encoding as we do for environment variables.

# Test Plan

Adjusted test. It was wrong all along! Passing of the rest of existing steps confirms steps work in general, right?