Support bash 3 in local builds
When testing an android app with maestro, the bash script that's generated relies the globstar option. This is only available on bash 4 or higher but mac os only comes with bash 3 (and will probably never move to bash 4 or 5 because they use the gplv3 license). Because of this if you run `eas -p android --local` on a mac the build will fail with this error (if the build runs maestro):

    shopt: globstar: invalid shell option name

To resolve this and support bash 3 I've refactored the maestro test code to use `find` rather than globstar, which should make it possible to run local builds on mac.