[build-tools] Allow specifying `deviceIdentifier` for `start_android_emulator` (#539)
# Why

https://exponent-internal.slack.com/archives/C06EFBQK3B7/p1748635317883789

# How

So for iOS we already have `device_identifier` which is "the device we want to use". Well, when Xcode is installed it autocreates a bunch of devices so the user just picks one or we default to shortest which is usually "iPhone XX".

Android comes with no virtual devices preset, so we create one. There is a `device_name` parameter which I have set a long time ago to `EasAndroidDevice01` and this is the name of the emulator we create.

However, we now need to have a way of providing a name of the device ID to use (`pixel_6` and others). So, I'm creating `device_template_id`. Or maybe it should be `device_identifier`, like it is on iOS?

# Test Plan

```yaml
jobs:
  hello_world:
    steps:
      - uses: eas/start_android_emulator
        with:
          device_template_id: pixel_6_pro
      - uses: eas/start_ios_simulator
        with:
          device_identifier: "iPad Pro 11-inch (M4)"
```

<img width="721" alt="Zrzut ekranu 2025-06-2 o 22 22 37" src="https://github.com/user-attachments/assets/4bbfee1e-2ae5-448f-a1e7-7b8440fe7bb4" />
<img width="1164" alt="Zrzut ekranu 2025-06-2 o 22 19 28" src="https://github.com/user-attachments/assets/81983732-3b06-4e44-bb84-6139e7f7a9d6" />
<img width="660" alt="Zrzut ekranu 2025-06-2 o 22 19 50" src="https://github.com/user-attachments/assets/c5a68622-ac6b-4594-b413-3503a37eba38" />