[build-tools] Refetch GitHub repository in worker before cloning (#541)
# Why

When we run jobs which use Git repository as project source, because the `repositoryUrl` contains an expiring token, `launcher` and `workflow-orchestration` need to call into `www` to refresh the `repositoryUrl`.

- The request is currently authenticated with a secret token which is bad — we could use `EXPO_TOKEN` to authenticate.
- Mike found this request often errors with 503 which causes problems.

# How

Before we clone the repository we're going to call `www` asking it for a fresh repository URL. If anything about it fails we're going to fall back to the original `repositoryUrl` which may still be valid. We're going to retry 3 times.

x-ref: https://github.com/expo/universe/pull/20501

# Test Plan

Tested end-to-end.

If data validation fails — we see the error logged:
<img width="1552" alt="Zrzut ekranu 2025-06-11 o 16 39 18" src="https://github.com/user-attachments/assets/4d823051-c61a-4f11-a3f8-ca1a4921a7de" />


If request succeeds — no "Prepare project" phase:

<img width="1165" alt="Zrzut ekranu 2025-06-11 o 16 39 29" src="https://github.com/user-attachments/assets/b616f02f-ddfc-4a94-8d95-cf6a57215a93" />