Move `env` from static interpolation context to dynamic (#550)
# Why

`env` should always be calculated in runtime from current values because it may change. It might also (in the future, once we land https://github.com/expo/universe/pull/19535) contain secrets and we're not encrypting this part of the job object.

# How

Moved `env` from static to dynamic interpolation context. This way we'll always remember to use a "live" value.

# Test Plan

Adjusted tests. TypeScript should ensure (and it did) that we won't forget to add `env: ...` wherever we're creating `JobInterpolationContext`.