[build-tools] bump repack (#553)
# Why

fixed the reentrant issue for repack on ios. when using fingerprint + repack for second round, if fingerprint doesn't change, the `get-build` will get the newer build, which is the repacked artifact. repack has an issue to deal with repacked app.

# How

the issue is resolved by https://github.com/expo/repack-app/pull/29 in 0.2.4

# Test Plan

tested repack workflow locally