[build-tools] Do not install Maestro if not required (#577)
# Why

We can speed up Maestro tests a bit if we use the preinstalled Maestro coming with the image. It's a new thing only present in `sdk-54` image, but still.

# How

Added a check that does not try to install Maestro if Maestro is present and no concrete `maestro_version` was requested by the developer.

If the developer requested `latest`, we install Maestro without providing a `MAESTRO_VERSION`. The respective part of the install script is:

```
if [ -z "$MAESTRO_VERSION" ]; then
    download_url="https://github.com/mobile-dev-inc/maestro/releases/latest/download/maestro.zip"
else
    download_url="https://github.com/mobile-dev-inc/maestro/releases/download/cli-$MAESTRO_VERSION/maestro.zip"
fi
```

so `latest` is obviously not a valid value.

# Test Plan

Added an integration test, it did not install Maestro for the first call, it did reinstall Maestro for the second call in the right version, it did reinstall Maestro with the latest version for the third call.