[chore] Upgrade TypeScript to 5.5 (#581)
# Why

I want to (after all) add https://gql-tada.0no.co/ to `build-tools`. Installation is simpler for TS 5.5+ so let's use this opportunity to ugprade to that.

# How

Upgraded TS to 5.5.4.

Did not upgrade to a more recent version not to have to deal with https://stackoverflow.com/questions/78790943/in-typescript-5-6-buffer-is-not-assignable-to-arraybufferview-or-uint8arr.

# Test Plan

Compared output of `yarn build` for current TS and new TS. The difference seems to be mostly about

<img width="830" height="895" alt="Zrzut ekranu 2025-08-26 o 12 41 24" src="https://github.com/user-attachments/assets/10c95b54-0adc-41f1-9518-84fbeb9bd2ac" />