[build-tools] Use hashed `${key}:${value}` for file secrets filename (#621)
# Why

Moving file secrets to a different directory did not help with fingerprint differences.

# How

This should. We're going to be get stable filenames for a given `(key, value)` pair. Moved secrets to working directory, in case the temporary path has some randomness to it too, like this part:

<img width="437" height="38" alt="Zrzut ekranu 2025-10-20 o 11 09 57" src="https://github.com/user-attachments/assets/6af51e8d-d639-4b17-b661-85f6d08442c9" />

# Test Plan

Added basic tests.