Fix local builds, export NPM tags (#624)
# Why

Fixing problems with recent code changes in `build-tools` when executing a local build. Adding `isLocal` and `isUsingCache` to the context, so checks for these are in one place and not duplicated throughout the code.

Related to https://github.com/expo/eas-cli/pull/3233, we would like to export the NPM tags for the EAS CLI version used in `build-tools` and in `www`, for consistency.  https://github.com/expo/eas-cli/pull/3237 will consume that to ensure that it is using the matching version of the local build plugin for local builds.

Finally, adding code in the local build plugin `main.ts` to print the error stack on failure, as otherwise any errors in `build-tools` are impossible to debug.


# Test Plan

CI should pass